% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_destination}
\alias{check_destination}
\title{check_destination}
\usage{
Checks locations of individuals in start against those in finish. Returns finish split into those that did and did not match their original start positions.
}
\arguments{
\item{start}{a population data frame with $X and $Y columns defining location.}

\item{finish}{a population data frame of the same individuals as start, with $X and $Y columns defining location.}
}
\value{
list with $matches as logical T/F. If F: $pop = finish. If T: $move_again = finish that didn't move and $the_rest = finsh that did move.
}
\description{
check_destination
}
\author{
Rose McKeon
}
