% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_germination}
\alias{check_germination}
\title{check_germination}
\usage{
check_germination(seeds, adults)
}
\arguments{
\item{seeds}{a population data frame with $X and $Y columns defining location, containing only seeds.}

\item{adults}{a population data frame with $X and $Y columns defining location, containing only adults.}
}
\value{
list with $matches as logical T/F. If F: $will_germinate = seeds. If T: $wont_germinate = seeds that are located where adults are established. $will_germinate = seeds that are located where no adults are established.
}
\description{
Checks locations of seeds against those of adults. Returns seeds split into those that could and could not germinate based on the presence of adults.
}
\author{
Rose McKeon
}
