% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{choose_alleles}
\alias{choose_alleles}
\title{choose_alleles}
\usage{
Chooses an allele, or N number of alleles from a locus.
}
\arguments{
\item{locus}{A numerical vector of allele values for the same locus. The genome of an individual which is grouped by locus and nested is passed to this function by sample_alleles using apply which makes every row of the genome in this format into a vector.}

\item{N}{integer representing the number of alleles to sample, ie: the ploidy level (default = 1 for haploid gametes).}
}
\value{
a random allele sampled from x.
}
\description{
choose_alleles
}
\author{
Rose McKeon
}
