% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compete}
\alias{compete}
\title{compete}
\usage{
Makes sure number of adults that survive in a cell to reproduce is constrained to K.
}
\arguments{
\item{competitors}{row of nested dataframe where competitors are grouped by location into list-column $plants.}

\item{K}{integer representing K, the carrying capacity (max population size) of any given cell (default = 1, so plants compete over grid squares and only 1 per square can survive).}
}
\value{
K winners sampled from competitors.
}
\description{
compete
}
\author{
Rose McKeon
}
