% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_genome}
\alias{create_genome}
\title{create_genome}
\usage{
Creates diploid genomes using random numbers.
}
\arguments{
\item{genome_size}{integer value for number of loci in genome}
}
\value{
Data frame containing a diploid genome, with 2 alleles for each locus and number of loci determined by genome_size. Data in long format with $allele and $locus as factors and $value being a random uniform number between 0 and 100.
}
\description{
create_genome
}
\examples{
create_genome()
}
\author{
Rose McKeon
}
