% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_pop}
\alias{create_pop}
\title{create_pop}
\usage{
Makes sure the population structure is always the same. Can be used to return an empty tibble (with no parameters supplied), or a generated population (if both parameters supplied).
}
\arguments{
\item{pop_size}{the number of rows to populate dataframe with.}

\item{grid_size}{the upper limit of landscape coordinates.}

\item{sim}{integer representing the current simulation number.}
}
\value{
Either an empty population dataframe with all required columns, or a filled population dataframe depending on supplied parameters.
}
\description{
create_pop
}
\examples{
create_pop()
create_pop(100, 10)
}
\author{
Rose McKeon
}
