% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_seeds}
\alias{create_seeds}
\title{create_seeds}
\usage{
Takes a dataframe containing zygote information (as created by create_zygotes) and returns a population dataframe in the format of that created by create_pop for which each zygote is now represented as a seed and has a genome which has been sampled from both parents, may have undergone genome duplication and may also have had alleles which have been mutated.
}
\arguments{
\item{zygotes}{dataframe of zygotes with progenitor IDs in $mum and $dad columns, as well as maternal location information in $X and $Y.}

\item{parents}{population dataframe containing all possible parents, complete with genetic information from which to sample.}

\item{generation}{integer used to prefix IDs of new individual seeds.}

\item{genome_size}{integer representing genome size of population.}

\item{mutation_rate}{number between 0 and 1 which represents the probability that any given allele will mutate.}
}
\value{
dataframe of paired ovules and pollen represented as progenitor IDs in columns $mum and $dad. Location data in columns $X and $Y maintained from ovule locations.
}
\description{
create_seeds
}
\author{
Rose McKeon
}
