% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_zygotes}
\alias{create_zygotes}
\title{create_zygotes}
\usage{
create_zygotes(mums = NULL, dads = NULL, ploidy_prob = 0.01,
  fertilisation_prob = 0.5, uneven_matching_prob = 0.1,
  selfing_polyploid_prob = 1, selfing_diploid_prob = 0,
  triploid_mum_prob = 0.1)
}
\arguments{
\item{mums}{dataframe of mums, adults with ovules already created using create_ovules().}

\item{dads}{dataframe of pollen donors (dads) adults who are within pollen_range of mums, found using get_pollen_donors().}

\item{ploidy_prob}{number between 0 and 1 which represents the probability that genome duplication will occur (default = 0.01).}

\item{fertilisation_prob}{number between 0 and 1 representing the probability that fertlisation will be successful (default = 0.5).}

\item{uneven_matching_prob}{number between 0 and 1 representing fertlisation_prob applied to zygotes with gametes whose ploidy levels do not match (default = 0.1 so triploids are rare but do occur).}

\item{selfing_polyploid_prob}{number between 0 and 1 representing fertilisation_prob applied to polyploids which are selfing (default = , so polyploids can always self)..}

\item{selfing_diploid_prob}{number between 0 and 1 representing fertilisation_prob applied to diploids which are selfing (default = 0, so diploids can never self).}

\item{triploid_mum_prob}{number between 0 and 1 representing fertilisation_prob applied to zygotes with triploid mums (default = 0.1 to reduce the number of seeds that triploid plants produce).}
}
\value{
dataframe of paired ovules and pollen represented as progenitor IDs in columns $mum and $dad. Location data in columns $X and $Y maintained from ovule locations.
}
\description{
Takes a dataframe containing an adult population that has ovules already created and returns a dataframe of successful fertilisations as progenitor ID pairs. Assumes all ovules receive pollen but applies a probability of fertlisation success to simulate some failure via mechanisms such as pollen loss etc.
}
\author{
Rose McKeon
}
