% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{disturb}
\alias{disturb}
\title{disturb}
\usage{
Causes mortality in a specific region of the landscape.
}
\arguments{
\item{pop}{a population dataframe as generated by populate_landscape.}

\item{prob}{a number between 0 and 1 representing probability of death.}

\item{xlim}{a vector of min and max X coordinates for the region where distrubance will occur.}

\item{grid_size}{an integer defining max grid coordinates for X or Y.}
}
\value{
pop with reductions based on location of individuals and probability of death.
}
\description{
disturb
}
\examples{
pop <- populate_landscape(); disturb(pop)
qplot(X, Y, data = pop, xlim = c(0, 100))
qplot(X, Y, data = disturb(pop), xlim = c(0, 100))
}
\author{
Rose McKeon
}
