% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fill_seeds_with_genomes}
\alias{fill_seeds_with_genomes}
\title{fill_seeds_with_genomes}
\usage{
Takes seed data (with linegae details nested in $genome) and replaces this column with real genomes; alleles sampled from parent genomes.
}
\arguments{
\item{seeds}{a data frame containing seeds.}

\item{parents}{a data frame containing possible parents.}

\item{mutation_rate}{number between 0 and 1 representing the rate of allele mutation (default = 0.001).}

\item{genome_size}{The size of the genome.}
}
\value{
seeds updated so $genome is replaced with real genomes.
}
\description{
fill_seeds_with_genomes
}
