% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{germinate}
\alias{germinate}
\title{germinate}
\usage{
Change life stage from seed to juvenile based on probability. New juveniles will have a size of 0.1.
}
\arguments{
\item{pop}{a population dataframe as generated by populate_landscape(), ie: nested by plant.}

\item{probability}{a number between 0 and 1 representing probability of germination.}
}
\value{
pop with updated life stages and sizes.
}
\description{
germinate
}
\author{
Rose McKeon
}
