% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{get_inbreeding_value}
\alias{get_inbreeding_value}
\title{get_inbreeding_values}
\usage{
Uses all alleles at the requested locus to determine whether or not a fiteness disadvantage due to inbreeding should be applied.
}
\arguments{
\item{genome}{A dataframe contining the genome of an individual}

\item{inbreeding_locus}{a positive integer which represents the locus used to detect inbreeding (default = 2).}
}
\description{
get_inbreeding_values
}
