% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_juveniles_not_outcompeted}
\alias{get_juveniles_not_outcompeted}
\title{Which Juveniles Survive Growth and Competition?}
\usage{
get_juveniles_not_outcompeted(adults, juveniles)
}
\arguments{
\item{adults}{a data frame containing adults.}

\item{juveniles}{a data frame containing juveniles.}
}
\value{
dataframe passed as juveniles modified with removals based on X and Y coordinates of adults.
}
\description{
Reduces juvenile population based on locations of adults. Juveniles cannot survive where adults become established.
}
\author{
Rose McKeon
}
