% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_those_not_outcompeted}
\alias{get_those_not_outcompeted}
\title{Which individuals are not challenged by adults?}
\usage{
get_those_not_outcompeted(adults, pop)
}
\arguments{
\item{adults}{a data frame containing adults.}

\item{pop}{data frame containing individuals, ie: seeds/juveniles.}

\item{message}{String containing log to output before giving number of removals (default = "Individuals outcompeted by adults: ").}
}
\value{
dataframe passed as inds modified with removals based on X and Y coordinates of adults.
}
\description{
Reduces population based on locations of adults. Used when those where adults become established cannot survive (100 percent mortality).
}
\author{
Rose McKeon
}
