% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_movement}
\alias{make_movement}
\title{make_movement}
\usage{
make_movement(pop, movement = NULL, grid_size = 100)
}
\arguments{
\item{pop}{population dataframe nested by plant}

\item{movement}{vector of adjustment values which will be added to X and Y values to give new coordinates. Values from movement are sampled at random to give directionless random movement.}

\item{grid_size}{integer, defines edges at which point movements wrap to the opposite side.}
}
\value{
pop with adjusted X and Y for every row.
}
\description{
updates coordinate variables of a population dataframe as generated by populate_landscape(), ie: should be nested by plant. Movement has wrapping boundaries depending on grid size.
}
\author{
Rose McKeon
}
