% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mutate_genome}
\alias{mutate_genome}
\title{mutate_genome}
\usage{
mutate_genome(genome = NULL, mutation_rate = 0.001)
}
\arguments{
\item{genome}{a dataframe containing the genome of an individual}

\item{mutation_rate}{a number representing the probability that mutation will occur. Used by rbinom to randomly choose which alleles are mutated.}
}
\value{
The genome data with allele values of mutated alleles updated if mutation occurred. Allele values are random uniform numbers between 0 and 100, chosen in the same way as a new allele in create_genome().
}
\description{
Mutate allele values of genome based on mutaion rate.
}
\author{
Rose McKeon
}
