% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nest_by_location}
\alias{nest_by_location}
\title{nest_by_location}
\usage{
Groups population dataframe by coordinates X and Y and nests all other data in list-column $plants. Arranged by X and Y.
}
\arguments{
\item{pop}{population data frame containing X and Y columns.}
}
\value{
updated pop.
}
\description{
nest_by_location
}
\author{
Rose McKeon
}
