% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reproduce}
\alias{reproduce}
\title{reproduce}
\usage{
Creates new seeds based on adult population.
}
\arguments{
\item{adults}{population dataframe that contains all adults of a particular generation.}

\item{N_ovules}{number of ovules each adult will create (default = 100).}

\item{pollen_range}{integer representing the dispersal rage of pollen default = 100 so, as grid_size default is also 100, all plants in the landscape will be used as potential pollen donors for all ovules. When < 100 only plants within range will be used as pollen donors, so alleles movement will be restricted into regions of the landscape.}

\item{fertilisation_prob}{number between 0 and 1 representing probability fertilisation between gametes is successful.}

\item{uneven_matching_prob}{number between 0 and 1 representing fertlisation_prob applied to zygotes with gametes whose ploidy levels do not match (default = 0.1 so triploids are rare but do occur).}

\item{selfing_polyploid_prob}{number between 0 and 1 representing fertilisation_prob applied to polyploids which are selfing (default = , so polyploids can always self)..}

\item{selfing_diploid_prob}{number between 0 and 1 representing fertilisation_prob applied to diploids which are selfing (default = 0, so diploids can never self).}

\item{triploid_mum_prob}{number between 0 and 1 representing fertilisation_prob applied to zygotes with triploid mums (default = 0.1 to reduce the number of seeds that triploid plants produce).}

\item{generation}{integer representing the generation of simulation in which offspring appear. Used to prefix individual $ID (default = 1).}

\item{ploidy_prob}{number between 0 and 1 representing the probability that genome duplication will occur (default = 0.01).}

\item{genome_size}{integer representing size of genome, ie: number of loci (default = 2).}

\item{grid_size}{integer representing the size of the landscape grid. Used to ensure pollen range does not exceed size of the landscape and reduces computation when pollen range is at maximum, ie: equal to grid size (default = 100).}
}
\description{
reproduce
}
\examples{
reproduce(adults)
}
\author{
Rose McKeon
}
