% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survive}
\alias{survive}
\title{survive}
\usage{
Reduces population data based on probability of survival.
}
\arguments{
\item{pop}{a population dataframe as generated by populate_landscape(), ie: nested by plant.}

\item{prob}{a number between 0 and 1 representing probability of survival (default = 1, all are very likely to survive).}

\item{inbreeding_sensitivity}{number between 0 and 1 representing the strength of inbreeding. 0 = no effect and 1 is maximum effect. Checking for 100% identical alleles at the specified inbreeding locus is used as a proxy for having homozygous deleterious alleles. When this happens survival chances are modified according to inbreeding sensitivity (default = 0, so inbreeding is ignored unless set).}
}
\value{
pop minus deaths
}
\description{
survive
}
\author{
Rose McKeon
}
