% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compete}
\alias{compete}
\title{compete}
\usage{
compete(competitors, K = 1, beneficial_trait = NULL)
}
\arguments{
\item{competitors}{row of nested dataframe where competitors are grouped by location into list-column $plants.}

\item{K}{integer representing K, the carrying capacity (max population size) of any given cell (default = 1, so plants compete over grid squares and only 1 per square can survive).}

\item{beneficial_trait}{the name of a column which contains trait values that should give a competative advatage, eg: "size".}
}
\value{
list where $adults = K winners sampled from competitors and $juveniles = juveniles with those in the same location as winnning adults removed.
}
\description{
Makes sure number of adults that survive in a cell to reproduce is constrained to K.
}
\author{
Rose McKeon
}
