% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_ovules}
\alias{create_ovules}
\title{create_ovules}
\usage{
create_ovules(plant, N = 500)
}
\arguments{
\item{plant}{A single row of a population dataframe as generated by populate_landscape. Should also be an adult (life stage == 2).}

\item{N}{integer representing the number of ovules to make.}
}
\value{
The updated plant dataframe.
}
\description{
Adds a new list-column $ovules to an individual plant which contains ovules as parent IDs.
}
\author{
Rose McKeon
}
