% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_zygotes_as_seeds}
\alias{format_zygotes_as_seeds}
\title{format_zygotes_as_seeds}
\usage{
Takes a dataframe containing zygote information (as created by create_zygotes) and returns them formatted as seeds. The result is a population dataframe with all the zygote information nested in the $genome list-column; ready to be filled with actual genomes at a later stage.
}
\arguments{
\item{zygotes}{dataframe of zygotes with progenitor IDs in $mum and $dad columns, as well as maternal location information in $X and $Y.}

\item{generation}{integer used to prefix IDs of new individual seeds.}
}
\value{
zygotes updated with $ID, $life_stage, $size and $genome (containing temporary lineage details used later to create real genomes).
}
\description{
format_zygotes_as_seeds
}
\author{
Rose McKeon
}
