% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{germinate}
\alias{germinate}
\title{germinate}
\usage{
germinate(seeds = NULL, adults = NULL, prob = 0.5)
}
\arguments{
\item{seeds}{a population dataframe that contains only seeds, as generated by populate_landscape(). These are the seeds that will germinate according to prob.}

\item{adults}{a population dataframe that contains only adults, as generated by populate_landscape(). If an adult is already established in a landscape cell, seeds in that cell will not germinate.}

\item{prob}{a number between 0 and 1 representing probability of germination.}
}
\value{
seeds with some individuals having updated $life_stage and $size, both of 1.
}
\description{
Changes some seeds within a dataframe into juveniles, based on the locations of adults as well as the probability of germination.
}
\author{
Rose McKeon
}
