% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pollen_donors}
\alias{get_pollen_donors}
\title{get_pollen_donors}
\usage{
get_pollen_donors(X = NULL, Y = NULL, adults = NULL,
  pollen_range = 99, grid_size = 100)
}
\arguments{
\item{X}{integer representing the X location of ovules.}

\item{Y}{integer representing the Y location of ovules.}

\item{adults}{dataframe of individual adults who may or may not be pollen donors.}

\item{pollen_range}{integer between 0 and grid_size - 1 representing the dispersal range of pollen (default = 99).}

\item{grid_size}{positive integer representing the size of the landscape grid (default = 100).}
}
\value{
dataframe subset of adults.
}
\description{
Filters a population dataframe by X and Y values within range of X and Y to give a subset of possible pollen donors for ovules that exist in X  and Y, based on pollen range and grid size (to account for landscape wrapping).
}
\examples{
# samples IDS of adults within 10 cell range of coordinate 1, 1.
get_pollen_donors(1, 1, adults, 10)
}
\author{
Rose McKeon
}
