% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grow}
\alias{grow}
\title{grow}
\usage{
Change size and number of plants based on growth rate determined via loci of genome.
}
\arguments{
\item{pop}{a population dataframe as generated by populate_landscape(), ie: nested by plant.}

\item{type}{character string defining type of growth. Can be "individuals" or "clones".}

\item{adult_size}{number representing size threshold for becoming an adult. Only used in individual growth, not clonal (default = NULL).}
}
\value{
pop with updated sizes and life stages.
}
\description{
grow
}
\author{
Rose McKeon
}
