% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{move}
\alias{move}
\title{move}
\usage{
updates coordinate variables of a population dataframe as generated by populate_landscape(), ie: should be nested by plant. Movement has wrapping boundaries depending on grid size.
}
\arguments{
\item{pop}{population dataframe nested by plant}

\item{grid_size}{integer, defines edges at which point movements wrap to the opposite side.}

\item{always_away}{logical value which determines if current location may or may not be used as final destination (default = FALSE, so it can be).}
}
\value{
pop with adjusted X and Y for every row.
}
\description{
move
}
\examples{
pop <- populate_landscape()
move(pop)
}
\author{
Rose McKeon
}
